+++ CHANGELOG.md
@@ -16,6 +16,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - update to dependencies to latest graphANNIS version
 
+### Fixed
+
+- Fix non-resolved relative path when importing EXMARaLDA files. 
+
 ## [0.7.0] - 2024-05-23
 
 ### Added